# swagger.yaml
openapi: 3.0.0
info:
  title: Code Vimarsh API
  version: 1.0.0
  description: Backend API documentation for Code Vimarsh – MSUB Coding Club
servers:
  - url: http://localhost:5000/api/v1
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
paths:
...
  /events:
    get:
      summary: Get all events
      tags:
        - Events
      responses:
        "200":
          description: List of all events
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  events:
                    type: array
                    items:
                      type: object
    post:
      summary: Create a new event
      tags:
        - Events
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - type
                - start_date
              properties:
                title:
                  type: string
                  minLength: 3
                  example: Introduction to Open Source
                description:
                  type: string
                  example: A beginner-friendly session on contributing to open source.
                long_description:
                  type: string
                  example: This workshop covers forking, branching, PRs, and more.
                type:
                  type: string
                  enum:
                    - Workshop
                    - Hackathon
                    - Meetup
                  example: Workshop
                location:
                  type: string
                  example: Lab 3, MSUB Campus
                start_date:
                  type: string
                  format: date-time
                  example: 2026-04-10T10:00:00.000Z
                end_date:
                  type: string
                  format: date-time
                  example: 2026-04-10T13:00:00.000Z
                max_participants:
                  type: integer
                  minimum: 1
                  example: 50
                banner_image:
                  type: string
                  format: uri
                  example: https://example.com/banner.png
                topics:
                  type: array
                  items:
                    type: string
                  default: []
                  example:
                    - Git
                    - GitHub
                    - Open Source
      responses:
        "201":
          description: Event created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  event:
                    type: object
        "401":
          description: Unauthorized
        "403":
          description: Forbidden – SUPER_ADMIN only
  /events/{id}:
    get:
      summary: Get an event by ID
      tags:
        - Events
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  event:
                    type: object
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Event not found
    put:
      summary: Update an event by ID
      tags:
        - Events
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 3
                  example: Updated Event Title
                description:
                  type: string
                  example: Updated description.
                long_description:
                  type: string
                  example: Updated long description.
                type:
                  type: string
                  enum:
                    - Workshop
                    - Hackathon
                    - Meetup
                  example: Hackathon
                location:
                  type: string
                  example: Auditorium, MSUB Campus
                start_date:
                  type: string
                  format: date-time
                  example: 2026-04-15T09:00:00.000Z
                end_date:
                  type: string
                  format: date-time
                  example: 2026-04-15T17:00:00.000Z
                max_participants:
                  type: integer
                  minimum: 1
                  example: 100
                banner_image:
                  type: string
                  format: uri
                  example: https://example.com/new-banner.png
                topics:
                  type: array
                  items:
                    type: string
                  example:
                    - DSA
                    - Web Dev
      responses:
        "200":
          description: Event updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  event:
                    type: object
        "401":
          description: Unauthorized
        "403":
          description: Forbidden – SUPER_ADMIN only
        "404":
          description: Event not found
    delete:
      summary: Delete an event by ID
      tags:
        - Events
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Event deleted successfully.
                  event:
                    type: object
        "401":
          description: Unauthorized
        "403":
          description: Forbidden – SUPER_ADMIN only
        "404":
          description: Event not found
  /events/{id}/register:
    post:
      summary: Register for an event
      tags:
        - Events
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - whatsapp_number
                - experience_level
              properties:
                whatsapp_number:
                  type: string
                  example: "9876543210"
                github_username:
                  type: string
                  example: johndoe
                experience_level:
                  type: string
                  enum:
                    - Beginner
                    - Intermediate
                    - Advanced
                  example: Beginner
                areas_of_interest:
                  type: array
                  items:
                    type: string
                  default: []
                  example:
                    - Web Dev
                    - AI/ML
                reason_for_participation:
                  type: string
                  example: I want to improve my competitive programming skills.
      responses:
        "201":
          description: Successfully registered for the event
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  registration:
                    type: object
                  message:
                    type: string
                    example: Successfully registered! A confirmation email has been sent.
        "401":
          description: Unauthorized
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Event not found
        "409":
          description: Already registered for this event
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: You are already registered for this event.
...